requirements.txt
google-api-python-client
oauth2client==4.1.2
nats-client==0.8.4
PyYAML==3.12
tornado==5.1.1